% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_6400.R
\name{read_6400}
\alias{read_6400}
\title{Reads data from LI-6400}
\usage{
read_6400(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
Returns a tibble from raw Licor 6400 files. Current
support for Licor 6400  wit soil CO2 flux chamber configuration.
}
\description{
Read data from the plain text or txt files and organizes it into
a tibble to add this data into R workflow.
}
\examples{
\dontrun{
read_6400("/path/filename")
}
}
